treadsRacing

+++ src/main/java/com/example/limittransactsapi/Model/RateDataFromJson.java
@@ -2,10 +2,11 @@ package com.example.limittransactsapi.Model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 
 @Data
 public class RateDataFromJson {
-    private Double closeRate;
+    private BigDecimal closeRate;
     private OffsetDateTime dateTime;
 }
